Implement traits on the ArchivedDecimal.
rkyv generates a separate type `ArchivedDecimal`, which represents
the serialized value. In this PR we implement several traits
to allow for ordering and equality traits on the ArchivedDecimal type
(according to the rules of the Decimal type).

We also implement comparison between `Decimal` and `ArchivedDecimal`
(PartialOrd) which is helpful to efficiently compare e.g., a more
complex in-memory type that uses decimal with it's serialized
counter-part like something that's stored on-disk.

Signed-off-by: Gerd Zellweger <mail@gerdzellweger.com>