Improved exp
Improved algorithm to compute exp (and thus also pow) to not compute
x^i directly in the Taylor series as it quickly outgrows exp(x).
Instead the summand is grown with x/i. exp can no be computed for
values up to ~66.5

Additionally the loop avoids two unnecessary ops for the tolerance
check and does not need to use the FACTORIAL table anymore. Overall
performance improved by about 5%.

As part of the new method the total iteration count of the algorithm
was increased to better meet the given tolerance. As runtime of exp
is linear in x this means that larger x will take to compute longer
than smaller x. This is expected and similar to how f64 works.