Improved exp
Improved algorithm to compute exp (and thus also pow) to not compute
x^i directly in the Taylor series as it quickly outgrows exp(x).
Instead the summand is grown with x/i. This also avoid using the
FACTORIAL table.

Additionally the loop avoids two unnecessary ops for the tolerance
check. In total performance improved by about 5%.

As part of the new method the total iteration count of the algorithm
was increased to better meet the given tolerance. This only affects
computation for large x. Bench test shows about 4x less performance
for x > ~18.